Fixed linux compiler issues broken by Android port
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@11509 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/src/test/os/OsLockTest.cpp
@@ -39,9 +39,9 @@ public:
         // Acquire semaphore at the start of the block, release it on exit block
         {
             int ret = pBSem->tryAcquire();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_SUCCESS, ret);
             ret = pBSem->release();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_SUCCESS, ret);
 
             OsLock lock(*pBSem);
             // if this were a real guarded method, we'd do useful work here
@@ -50,7 +50,7 @@ public:
 
             // Should not be available
             ret = pBSem->tryAcquire();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_BUSY, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_BUSY, ret);
         }
 
         delete pBSem;
@@ -69,70 +69,70 @@ public:
         // Acquire read lock at the start of the block, release it on exit
         {
             ret = pRWMutex->tryAcquireRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
 
             // Multiple read should be allowed
             ret = pRWMutex->tryAcquireRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
 
             // Should not be available
             ret = pRWMutex->tryAcquireWrite();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should not be available", OS_BUSY, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should not be available", (int)OS_BUSY, ret);
 
             ret = pRWMutex->releaseRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_SUCCESS, ret);
             ret = pRWMutex->releaseRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_SUCCESS, ret);
 
             OsReadLock lock(*pRWMutex);
 
             // Should not be available
             ret = pRWMutex->tryAcquireWrite();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should not be available", OS_BUSY, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should not be available", (int)OS_BUSY, ret);
 
             // Multiple read should be allowed
             ret = pRWMutex->tryAcquireRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
             ret = pRWMutex->releaseRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_SUCCESS, ret);
 
         }
 
         // Should be available
         ret = pRWMutex->tryAcquireWrite();
-        CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+        CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
         ret = pRWMutex->releaseWrite();
-        CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", OS_SUCCESS, ret);
+        CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", (int)OS_SUCCESS, ret);
 
         // Acquire write lock at the start of the block, release it on exit
         {
             ret = pRWMutex->tryAcquireRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
             ret = pRWMutex->releaseRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", (int)OS_SUCCESS, ret);
 
 
             // Should be available
             ret = pRWMutex->tryAcquireWrite();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
             ret = pRWMutex->releaseWrite();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", OS_SUCCESS, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", (int)OS_SUCCESS, ret);
 
 
             OsWriteLock lock(*pRWMutex);
 
             // Should not be available
             ret = pRWMutex->tryAcquireWrite();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_BUSY, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_BUSY, ret);
             ret = pRWMutex->tryAcquireRead();
-            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", OS_BUSY, ret);
+            CPPUNIT_ASSERT_EQUAL_MESSAGE("Semaphor should be available", (int)OS_BUSY, ret);
         }
 
         // Should be available
         ret = pRWMutex->tryAcquireWrite();
-        CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", OS_SUCCESS, ret);
+        CPPUNIT_ASSERT_EQUAL_MESSAGE("should be available", (int)OS_SUCCESS, ret);
         ret = pRWMutex->releaseWrite();
-        CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", OS_SUCCESS, ret);
+        CPPUNIT_ASSERT_EQUAL_MESSAGE("should release", (int)OS_SUCCESS, ret);
 
         delete pRWMutex;
     }
